Add feature to check for internet.
Tools can specify that they need a internet connection.
If so, it will be checked before tool gets executed.
The execution is skipped, if the requirement is not fulfilled.
Additionally the interval of the tools handler will be adjusted while no
internet is available to check more frequently if the requirement is
present.